requirements.txt
streamlit==1.32.0
ultralytics==8.2.0
tensorflow==2.15.0
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.3.0